Reset connection address to one of the cluster nodes ...
after testing that the added Virtual IP is working.

Currently, after test_add_virtual_ip has added a Virtual IP to the
cluster, test script will connect to HAWK via the Virtual IP to
verify that the address was successfully configured in the cluster. On
success, test script will reset the connection address to the one
originally supplied in the -H/--host option; however, if connection to HAWK
via the Virtual IP is not possible, connection address is not reset and
all remaining tests will fail to connect to HAWK.

This commit resets the connection address even on failures, and also
rewords the error message to make it more clear.